+++ src/util/serialization.js
@@ -3,12 +3,22 @@
 // Modified by zelthrStudio (2026) from the original `request` package
 // (Copyright 2010-2012 Mikeal Rogers, Apache License 2.0).
 
+// Headers never serialized: logging a request/response object must not leak
+// credentials (Authorization/Bearer), cookies or API keys.
+const SENSITIVE_HEADER_RE = /^(authorization|proxy-authorization|cookie|set-cookie|x-api-key|api-key|apikey)$/i
+
 // Return a simpler request object to allow serialization.
 function requestToJSON () {
+  const headers = {}
+  for (const key of Object.keys(this.headers || {})) {
+    if (!SENSITIVE_HEADER_RE.test(key)) {
+      headers[key] = this.headers[key]
+    }
+  }
   return {
     uri: this.uri,
     method: this.method,
-    headers: this.headers
+    headers
   }
 }